minor tweaks

+++ task1-C/main.cpp
@@ -5,7 +5,7 @@ int main() {
     std::string pattern, text;
     std::cin >> pattern >> text;
     auto indices = FindPattern(pattern, text);
-    for (int i : indices)
+    for (size_t i : indices)
         std::cout << i << ' ';
     return 0;
 }
