+++ task1-C/string_functions.hpp
@@ -2,7 +2,7 @@
 #include <vector>
 #include <string>
 
-std::vector<int> FindPattern(const std::string& pattern,
+std::vector<size_t> FindPattern(const std::string& pattern,
     const std::string& text);
 
 std::vector<std::string> Split(const std::string& str, char delimeter);
